requirements.txt
Django>=5.0
requests>=2.32.0
python-dotenv>=1.0.1
